Use URLs.join instead of File.join to compose URLs
File.join uses File::SEPARATOR to join paths, which is not correct for
URLs. This commit changes code that joins URL parts to use URLs.join
instead. This method always uses "/" as a separator.

An alternative I considered was to use URI.join, but it's API is not
really suitable for our use cases.